Move files into the dlsoo package. Fix plumbing.
