+++ dlsoo/CONFIGURATIONS/Simulation_Config
@@ -8,7 +8,7 @@ p3
 (dp4
 S'mr_obj'
 p5
-(idls_optimiser_util
+(idlsoo.util
 dls_measurement_var
 p6
 (dp7
@@ -50,7 +50,7 @@ mr_representation
 p23
 (dp24
 g5
-(idls_optimiser_util
+(idlsoo.util
 dls_measurement_var
 p25
 (dp26
@@ -108,7 +108,7 @@ p44
 g38
 sS'mp_obj'
 p45
-(idls_optimiser_util
+(idlsoo.util
 dls_param_var
 p46
 (dp47
@@ -146,7 +146,7 @@ p56
 sg44
 g52
 sg45
-(idls_optimiser_util
+(idlsoo.util
 dls_param_var
 p57
 (dp58
@@ -182,7 +182,7 @@ p65
 sg44
 g61
 sg45
-(idls_optimiser_util
+(idlsoo.util
 dls_param_var
 p66
 (dp67
