+++ dlsoo/dlsoo_rcds.py
@@ -18,7 +18,7 @@ import Tkinter
 import ttk
 import tkMessageBox
 
-import dls_optimiser_plot as plot
+import .plot
 import matplotlib
 matplotlib.use("TkAgg")
 from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg, NavigationToolbar2TkAgg
